feat: rabbit service 구현

+++ src/main/java/shop/s5g/front/dto/coupon/coupon/AvailableCouponResponseDto.java
@@ -20,7 +20,7 @@ public record AvailableCouponResponseDto(
 
     BigDecimal discountPrice,
 
-    Integer count
+    Long count
 ) {
 
 }
